fix : Remove menu from MainActivity (#1940)
