--- android/app/src/main/res/menu/menu_extras.xml
@@ -1,6 +1,0 @@
-<?xml version="1.0" encoding="utf-8"?>
-<menu xmlns:android="http://schemas.android.com/apk/res/android">
-    <item
-        android:title="Copyright"
-        android:id="@+id/action_display_copyright_dialog"/>
-</menu>
