+++ android/app/src/main/res/menu/menu_home.xml
@@ -9,4 +9,8 @@
         android:icon="@drawable/ic_search_white_24dp"
         android:title="@string/search"
         app:showAsAction="always" />
+
+    <item
+        android:title="Copyright"
+        android:id="@+id/action_display_copyright_dialog"/>
 </menu>
